pyproject.toml
[tool.poetry]
name = "hw"
version = "0.1.0"
description = ""
authors = ["d-vasily <dvs15m315@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
catboost = "^1.2.2"
pandas = "^2.1.4"
black = "^23.12.1"
dvc = "^3.36.1"
mlflow = "^2.9.2"
pre-commit = "^3.6.0"
flake8 = "^6.1.0"
isort = "^5.13.2"
hydra-core = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
